Making the navigation work
I made the navigation work by fixing a few issues. Firstly, I had to change my object to something that actually works. Made away with unnecessary notation in the object.

Then I had to google and realize that I gotta use v-bind to make links work inside a v-for.

Not it's all good